Hang

+++ code/cpp/utility/utility.cpp
@@ -58,7 +58,7 @@ namespace IMUProject{
 	}
 
 	void TrajectoryOverlay(const double pixel_length, const Eigen::Vector2d& sp, const Eigen::Vector3d& map_ori,
-	                       const std::vector<Eigen::Vector3d>& positions, const Eigen::Vector3i& color, cv::Mat& map){
+	                       const std::vector<Eigen::Vector3d>& positions, const Eigen::Vector3d& color, cv::Mat& map){
 		CHECK(map.data);
 
 		// Assume
